requirements.txt
Flask==3.0.0
pytest==7.4.0
openai==1.54.4
python-dotenv==1.0.1
mock==5.1.0
ipytest==0.14.2
pytest-mock==3.14.0
